+++ bytecode/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.12.0.dev"
+__version__ = "0.12.0"
 
 __all__ = [
     "Label",