+++ doc/changelog.rst
@@ -18,6 +18,8 @@ Bugfixes:
 - Fix a design flaw in the flag inference mechanism that could very easily
   lead to invalid flags configuration PR #56
 
+- Fix recursion limitations when compiling bytecode with numerous basic
+  blocks. #57
 
 2020-02-02: Version 0.10.0
 --------------------------