+++ doc/changelog.rst
@@ -4,6 +4,10 @@ ChangeLog
 unreleased: Version 0.12.0
 --------------------------
 
+Bugfixes:
+
+- Fix recursion limitations when compiling bytecode with numerous basic
+  blocks. #57
 
 2020-03-02: Version 0.11.0
 --------------------------
@@ -17,9 +21,6 @@ Bugfixes:
 
 - Fix a design flaw in the flag inference mechanism that could very easily
   lead to invalid flags configuration PR #56
-
-- Fix recursion limitations when compiling bytecode with numerous basic
-  blocks. #57
 
 2020-02-02: Version 0.10.0
 --------------------------