+++ doc/changelog.rst
@@ -9,6 +9,7 @@ Bugfixes:
 - Fix recursion limitations when compiling bytecode with numerous basic
   blocks. #57
 
+
 2020-03-02: Version 0.11.0
 --------------------------
 
@@ -21,6 +22,7 @@ Bugfixes:
 
 - Fix a design flaw in the flag inference mechanism that could very easily
   lead to invalid flags configuration PR #56
+
 
 2020-02-02: Version 0.10.0
 --------------------------